+++ src/main/kotlin/com/pestphp/pest/features/uses/UsesReferenceContributor.kt
@@ -1,0 +1,21 @@
+package com.pestphp.pest.features.uses
+
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.psi.PsiReferenceContributor
+import com.intellij.psi.PsiReferenceRegistrar
+import com.jetbrains.php.lang.psi.elements.ParameterList
+import com.jetbrains.php.lang.psi.elements.StringLiteralExpression
+import com.jetbrains.php.lang.psi.elements.impl.MethodReferenceImpl
+
+class UsesReferenceContributor : PsiReferenceContributor() {
+    override fun registerReferenceProviders(registrar: PsiReferenceRegistrar) {
+        registrar.registerReferenceProvider(
+            PlatformPatterns.psiElement(StringLiteralExpression::class.java)
+                .withParents(
+                    ParameterList::class.java,
+                    MethodReferenceImpl::class.java
+                ),
+            UsesInDirectoryReferenceProvider()
+        )
+    }
+}