+++ src/test/resources/com/pestphp/pest/features/uses/CompleteFakeInFolder.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Tests;
+
+notUses(TestCase::class)->in('<caret>');