+++ src/test/resources/com/pestphp/pest/features/uses/CompleteInFolder.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Tests;
+
+uses(TestCase::class)->in('<caret>');